wip: render custom control based on schema matching
This example shows how custom schema extensions can be matched and
used to render a custom control. For the sake of example, a builtin
"MaterialBooleanControl" is rendered, but in practice this would be
our own custom control component, which would have access to the
custom `schema` passed via props.

To test: `npm run dev` and navigate to this URL

http://localhost:3000/?schema=N4KABGBEAOBOD20CmsAuBLJBnSAuMoEEkAtgJ4C0JAhtBQGaYA2AJngeEcamcu5FlSx0AOwDmkADScukevFgksAOhEAjMsvgikAegDGAV0HwSFfdqHwm7QlyKQeffAIAW1FkhYVUieACssbSkZe0c-QOD8O3sHIMNYfSQAYW0dfQwoqAsRdMyRChFqEiQKAHd0VFdRcyZ4Q29aOhEcUK4AX2lY4hZqVGpbNrD4xJS0pAz0LMgcvKmCopLyyuqC-TqGiibC1u6wdqGDjs4D9qA&uiSchema=N4KABBYEQC4J4AcCmUBc0BqSBOMCWAxgIYA2AMkXAPYCuMUANOJFEiUgLZIB2MAzmjABtZpEigxklvGSCoAYSq9sVEo1FSWfAlVnooAYgD0CFclx4kfIxzgBaDkQR2AZpZIATdZslRd%2BJQF0YABfDTEwsQBdEBCgA