Add dummy In-Line formatter example
This example adds a new data form, that breaks all the scrolling but
lets the user input the data.

With this, and the new inline control for dropdowns, it demonstrates how
to get dropDown values from values embedded in the JSON Data input.

This is to cover the scenario where SIM objects contain a list (dropdown)
of slice IDs is able to connect to. This IDs are per/SIM and part of the
CRD, so they cannot be added as pàrt of the block logic or the
cardinality would explode

Signed-off-by: OscarMoya <oscar.moya@nearbycomputing.com>